Start timers for newly enabled reminders
